Review fixes: no-controller guard, input isActive checks, spawn regen gate, test reorganization
- Split no-controller into separate `return true` guard matching official engine
- Add checkIsActive on input labs (runReaction, reverseReaction) and link target
- Gate spawn energy regen on isActive to prevent energyAvailable drift
- Remove unused checkIsActive import from mineral.ts
- Move generic isActive tests to structure/test.ts, observer tests stay in observer/test.ts
- Register structure mod test in manifest
- Rename spawn describe block and remove redundant assertion
