+++ src/mods/logistics/link.ts
@@ -68,6 +68,7 @@ export function checkTransferEnergy(link: StructureLink, target: StructureLink, 
 		() => checkMyStructure(link, StructureLink),
 		() => checkIsActive(link),
 		() => checkTarget(target, StructureLink),
+		() => checkIsActive(target),
 		() => checkHasResource(link, C.RESOURCE_ENERGY, amount),
 		() => checkHasCapacity(target, C.RESOURCE_ENERGY, amount),
 		() => checkSameRoom(link, target),
