+++ src/mods/structure/index.ts
@@ -1,5 +1,5 @@
 import type { Manifest } from 'xxscreeps/config/mods/index.js';
 
 export const manifest: Manifest = {
-	provides: [ 'backend', 'constants', 'game', 'processor' ],
+	provides: [ 'backend', 'constants', 'game', 'processor', 'test' ],
 };
